+++ tests/TestCase/Event/EventTest.php
@@ -16,7 +16,6 @@ declare(strict_types=1);
 namespace Tools\Test\Event;
 
 use Tools\Event\Event;
-use Tools\Exception\KeyNotExistsException;
 use Tools\TestSuite\TestCase;
 
 /**
@@ -57,7 +56,6 @@ class EventTest extends TestCase
         $this->assertSame('arg1', $this->Event->getArg(0));
         $this->assertSame('arg2', $this->Event->getArg(1));
 
-        $this->expectException(KeyNotExistsException::class);
         $this->expectExceptionMessage('Argument with index `2` does not exist');
         $this->assertSame('arg2', $this->Event->getArg(2));
     }